update duplicate handling spec for moveGranules to use IngestGranule workflow
