+++ example/app/config.yml
@@ -131,7 +131,6 @@ default:
     'workflows/errorsAndRetries.yml',
     'workflows/sips.yml',
     'workflows/syncGranule.yml',
-    'workflows/moveGranules.yml'
   ]
 
   iams:
