--- example/workflows/moveGranules.yml
@@ -1,78 +1,0 @@
-MoveGranules:
-  Comment: 'Move Granules'
-  StartAt: Report
-  States:
-    Report:
-      Type: Task
-      Resource: ${SfSnsReportLambdaFunction.Arn}
-      CumulusConfig:
-        cumulus_message:
-          input: '{$}'
-      Next: SyncGranuleNoVpc
-    SyncGranuleNoVpc:
-      CumulusConfig:
-        buckets: '{$.meta.buckets}'
-        provider: '{$.meta.provider}'
-        collection: '{$.meta.collection}'
-        stack: '{$.meta.stack}'
-        downloadBucket: '{$.cumulus_meta.system_bucket}'
-        pdr: '{$.meta.pdr}'
-        cumulus_message:
-          outputs:
-            - source: '{$.granules}'
-              destination: '{$.meta.input_granules}'
-            - source: '{$}'
-              destination: '{$.payload}'
-            - source: '{$.process}'
-              destination: '{$.cumulus_meta.process}'
-      Type: Task
-      Resource: ${SyncGranuleNoVpcLambdaFunction.Arn}
-      Catch:
-        - &ErrorEqualDefaults
-          ErrorEquals:
-          - States.ALL
-          ResultPath: '$.exception'
-          Next: StopStatus
-      Next: ProcessingStep
-    # Had to add this to transform payload into what MoveGranules task is expecting
-    ProcessingStep:
-      CumulusConfig:
-        bucket: '{$.meta.buckets.internal.name}'
-        collection: '{$.meta.collection}'
-      Type: Task
-      Resource: ${FakeProcessingLambdaFunction.Arn}
-      Catch:
-        - <<: *ErrorEqualDefaults
-      Next: MoveGranuleStep
-    MoveGranuleStep:
-      CumulusConfig:
-        bucket: '{$.meta.buckets.internal.name}'
-        buckets: '{$.meta.buckets}'
-        granuleIdExtraction: '{$.meta.collection.granuleIdExtraction}'
-        distribution_endpoint: '{$.meta.distribution_endpoint}'
-        input_granules: '{$.meta.input_granules}'
-        collection: '{$.meta.collection}'
-      Type: Task
-      Resource: ${MoveGranulesLambdaFunction.Arn}
-      Catch:
-        - <<: *ErrorEqualDefaults
-      Next: StopStatus
-    StopStatus:
-      Type: Task
-      Resource: ${SfSnsReportLambdaFunction.Arn}
-      CumulusConfig:
-        sfnEnd: true
-        stack: '{$.meta.stack}'
-        bucket: '{$.meta.buckets.internal.name}'
-        stateMachine: '{$.cumulus_meta.state_machine}'
-        executionName: '{$.cumulus_meta.execution_name}'
-        cumulus_message:
-          input: '{$}'
-      Catch:
-        - ErrorEquals:
-          - States.ALL
-          Next: WorkflowFailed
-      End: true
-    WorkflowFailed:
-      Type: Fail
-      Cause: 'Workflow failed'
